flask_template/requirements.txt
Flask
ibm_cloud_sdk_core
python-dotenv
gunicorn
multiprocess
Pillow
pytesseract
pdf2image
flask-jwt-extended
flask-cors
python-dotenv
pymongo
simplejson
ibmcloudant
ibm_cloud_sdk_core
datetime
flask_jwt_extended
boto3
neo4j
py2neo
unittest
pytest